tests: move tensorflow tests to their own file
Move tensorflow tests into separate `test_tensorflow` file.
Improve the `tensorflow_onedir_only` test mark (force pyi_builder
into generating only onedir case instead of skipping the onefile
case).
